refacto(888): extract ActoinCommentaireVM
